fix: corriger les 4 problemes bloquants identifies lors de l'audit backend
1. Race condition sur le solde (debit/credit sans verrou) : ajout de
   SELECT ... FOR UPDATE via un parametre for_update sur
   obtenir_compte_du_client(), utilise partout ou un compte va etre
   debite/credite. Pour les transferts, verrouillage dans un ordre
   constant (id_compte croissant, pas source->destination) pour eviter
   les deadlocks entre deux virements croises concurrents. Prouve par
   un test reel avec deux threads contre Postgres (le second thread
   attend 1.2s au lieu d'acceder immediatement).

2. Float pour l'argent : toutes les colonnes montant/solde/prix passent
   de Float a Numeric(14,2), y compris dans les modules dormants (plans,
   budgets, epargne, dettes, analyse). Les schemas Pydantic passent de
   float a Decimal en consequence (l'API renvoie desormais les montants
   en chaine "1000.00", pas en nombre flottant - plus sur cote client JS,
   comportement par defaut de Pydantic v2 pour Decimal).

3. Aucune contrainte DB empechant un solde negatif : ajout de
   CheckConstraint (solde >= 0 sur comptes_financiers, montant > 0 sur
   transactions et transferts), non detectees par l'autogenerate
   d'Alembic (limitation connue) donc ajoutees a la main dans la
   migration.

4. annuler_transaction() ne verifiait pas le decouvert resultant :
   calcule desormais l'impact de l'annulation AVANT d'inserer quoi que
   ce soit, et refuse (SoldeInsuffisantError) si le solde deviendrait
   negatif - scenario reel : annuler un DEPOT_INITIAL apres avoir
   depense une partie de l'argent.

Corrige au passage une incoherence de nommage relevee dans l'audit :
CompteIntrouvableError etait reutilisee a tort pour signaler une
transaction introuvable dans annuler_transaction ; nouvelle exception
TransactionIntrouvableError dediee.

5 nouveaux tests dans test_integrite_financiere.py (contraintes CHECK,
refus de decouvert via annulation, verrouillage FOR UPDATE). 34/34 tests
au total, verifie egalement en conditions reelles contre Postgres.

+++ backend/app/modules/analyse/models.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from sqlalchemy import Column, Integer, String, Float, Date, DateTime, JSON, ForeignKey
+from sqlalchemy import Column, Integer, String, Float, Numeric, Date, DateTime, JSON, ForeignKey
 from sqlalchemy.orm import relationship
 from app.core.database import Base
 
@@ -35,7 +35,7 @@ class Prediction(Base):
     type = Column(String, nullable=False)  # DEPENSES_FUTURES, RISQUE_BUDGETAIRE, CAPACITE_EPARGNE
     periode_debut = Column(Date, nullable=False)
     periode_fin = Column(Date, nullable=False)
-    montant_predit = Column(Float, nullable=True)
+    montant_predit = Column(Numeric(14, 2), nullable=True)
     niveau_confiance = Column(Float, nullable=True)
     recommandations = Column(String, nullable=True)
     date_generation = Column(DateTime, default=datetime.utcnow)
